refactor: admin challenge query separation
