+++ src/main/java/com/example/green/domain/challenge/controller/query/docs/AdminTeamChallengeQueryControllerDocs.java
@@ -1,10 +1,9 @@
-package com.example.green.domain.challenge.controller.docs;
+package com.example.green.domain.challenge.controller.query.docs;
 
 import static io.swagger.v3.oas.annotations.enums.ParameterIn.*;
 
 import com.example.green.domain.challenge.controller.dto.admin.AdminChallengeDetailDto;
 import com.example.green.domain.challenge.controller.dto.admin.AdminChallengeParticipantListResponseDto;
-import com.example.green.domain.challenge.controller.dto.admin.AdminPersonalChallengesDto;
 import com.example.green.domain.challenge.controller.dto.admin.AdminTeamChallengeGroupDetailResponseDto;
 import com.example.green.domain.challenge.controller.dto.admin.AdminTeamChallengeGroupListResponseDto;
 import com.example.green.domain.challenge.controller.dto.admin.AdminTeamChallengesDto;
@@ -21,17 +20,7 @@ import io.swagger.v3.oas.annotations.responses.ApiResponse;
 import io.swagger.v3.oas.annotations.tags.Tag;
 
 @Tag(name = "관리자 챌린지 관리 API", description = "관리자 챌린지 생성/수정/이미지/전시여부 등 관리 API")
-public interface AdminChallengeQueryControllerDocs {
-
-	@Operation(summary = "개인 챌린지 목록 조회", description = "개인 챌린지 목록을 조회합니다. (10개씩 조회)")
-	@ApiErrorStandard
-	@ApiResponse(responseCode = "200", description = "개인 챌린지 목록 조회 성공", useReturnTypeSchema = true)
-	@ApiResponse(responseCode = "403", description = "관리자 권한이 필요합니다.",
-		content = @Content(schema = @Schema(implementation = ExceptionResponse.class)))
-	ApiTemplate<CursorTemplate<Long, AdminPersonalChallengesDto>> getPersonalChallenges(
-		@Parameter(description = "커서 (마지막 챌린지 ID) - 첫 번째 조회 시에는 아무것도 넣지 말고 조회하세요") Long cursor,
-		@Parameter(description = "페이지 사이즈(생략 가능)", example = "20") Integer size
-	);
+public interface AdminTeamChallengeQueryControllerDocs {
 
 	@Operation(summary = "팀 챌린지 목록 조회", description = "팀 챌린지 목록을 조회합니다. (10개씩 조회)")
 	@ApiErrorStandard
@@ -42,15 +31,6 @@ public interface AdminChallengeQueryControllerDocs {
 		@Parameter(description = "커서 (마지막 챌린지 ID) - 첫 번째 조회 시에는 아무것도 넣지 말고 조회하세요") Long cursor,
 		@Parameter(description = "페이지 사이즈(생략 가능)", example = "20") Integer size
 	);
-
-	@Operation(summary = "개인 챌린지 상세 조회", description = "개인 챌린지 상세 정보를 조회합니다.")
-	@ApiErrorStandard
-	@ApiResponse(responseCode = "200", description = "챌린지 상세 조회 성공", useReturnTypeSchema = true)
-	@ApiResponse(responseCode = "403", description = "관리자 권한이 필요합니다.",
-		content = @Content(schema = @Schema(implementation = ExceptionResponse.class)))
-	ApiTemplate<AdminChallengeDetailDto> getPersonalChallengeDetail(
-		@Parameter(name = "challengeId", description = "챌린지 ID", in = PATH, required = true, example = "1")
-		Long challengeId);
 
 	@Operation(summary = "팀 챌린지 상세 조회", description = "팀 챌린지 상세 정보를 조회합니다.")
 	@ApiErrorStandard
